jira-pulsar-integration/requirements.txt
# pip install -r requirements.txt
requests
pulsar-client
python-dotenv
Flask
